Moving some function locations
