+++ sudokuru/app/Api/Puzzles.ts
@@ -3,29 +3,10 @@ import {
   SudokuObjectProps,
 } from "../Functions/LocalDatabase";
 import { getKeyJSON, removeData, storeData } from "../Functions/AsyncStorage";
-import {
-  GameDifficulty,
-  returnGameOfDifficulty,
-} from "../Components/SudokuBoard/Core/Functions/DifficultyFunctions";
+
 import { Statistics } from "./Puzzle.Types";
 import { getStatistics, saveStatisitics } from "./Statistics";
 import { SudokuStrategy } from "sudokuru";
-
-/**
- * Given a difficulty and an user auth token retrieves a random puzzle close to the difficulty that the user hasn't solved before
- * @param difficulty - difficulty number (between 0 and 1)
- * @param strategies - new game can have subset of these strategies
- * @param initializeNotes - boolean to determine if notes should be initialized.
- * @returns promise of puzzle JSON object
- */
-export const startGame = (
-  difficulty: GameDifficulty,
-  initializeNotes: boolean,
-): SudokuObjectProps => {
-  return returnGameOfDifficulty(difficulty, initializeNotes);
-  // !uncomment below for dev testing
-  // return returnGameOfDifficulty("dev", initializeNotes);
-};
 
 /**
  * Given an user auth token retrieves the users active game or returns null if the user doesn't have an active game
