+++ sudokuru/app/Components/SudokuBoard/Core/Functions/DifficultyFunctions.ts
@@ -1,18 +1,3 @@
-import { AMATEUR_PUZZLES } from "../../../../Data/puzzles/amateur_puzzles";
-import { GRANDMASTER_PUZZLES } from "../../../../Data/puzzles/grandmaster_puzzles";
-import { LAYMAN_PUZZLES } from "../../../../Data/puzzles/layman_puzzles";
-import { MASTER_PUZZLES } from "../../../../Data/puzzles/master_puzzles";
-import { NOVICE_PUZZLES } from "../../../../Data/puzzles/novice_puzzles";
-import { PROFESSIONAL_PUZZLES } from "../../../../Data/puzzles/professional_puzzles";
-import { PROTEGE_PUZZLES } from "../../../../Data/puzzles/protege_puzzles";
-import { PUNDIT_PUZZLES } from "../../../../Data/puzzles/pundit_puzzles";
-import { TRAINEE_PUZZLES } from "../../../../Data/puzzles/trainee_puzzles";
-import {
-  convertPuzzleToSudokuObject,
-  InputPuzzle,
-  SudokuObjectProps,
-} from "../../../../Functions/LocalDatabase";
-
 export type GameDifficulty =
   | "novice"
   | "amateur"
@@ -115,62 +100,3 @@ export function calculateGameScore(
   const timeScore: number = calculateTimeScore(time);
   return difficultyScore + hintAndIncorrectCellsScore + timeScore;
 }
-
-/**
- * Returns a random puzzle from an array of puzzles
- * @param PUZZLES An array of puzzles
- * @returns a random puzzle from an array of puzzles
- */
-const retrieveRandomPuzzle = (PUZZLES: InputPuzzle[]): InputPuzzle => {
-  return PUZZLES[Math.floor(Math.random() * PUZZLES.length)];
-};
-
-/**
- * This function takes in the requested difficulty and returns a puzzle matching the difficulty.
- * @param difficulty The difficulty classification of the puzzle. (string)
- * @returns A Puzzle object
- */
-export const returnPuzzleOfDifficulty = (
-  difficulty: GameDifficulty | "dev",
-): InputPuzzle => {
-  switch (difficulty) {
-    // "dev" difficulty is a custom difficulty that always returns the same puzzle.
-    case "dev":
-      return NOVICE_PUZZLES[0];
-    case "novice":
-      return retrieveRandomPuzzle(NOVICE_PUZZLES);
-    case "amateur":
-      return retrieveRandomPuzzle(AMATEUR_PUZZLES);
-    case "layman":
-      return retrieveRandomPuzzle(LAYMAN_PUZZLES);
-    case "trainee":
-      return retrieveRandomPuzzle(TRAINEE_PUZZLES);
-    case "protege":
-      return retrieveRandomPuzzle(PROTEGE_PUZZLES);
-    case "professional":
-      return retrieveRandomPuzzle(PROFESSIONAL_PUZZLES);
-    case "pundit":
-      return retrieveRandomPuzzle(PUNDIT_PUZZLES);
-    case "master":
-      return retrieveRandomPuzzle(MASTER_PUZZLES);
-    case "grandmaster":
-      return retrieveRandomPuzzle(GRANDMASTER_PUZZLES);
-  }
-};
-
-/**
- * This function takes in the requested difficulty and returns a puzzle matching the difficulty.
- * @param difficulty The difficulty classification of the puzzle. (string)
- * @param initializeNotes Boolean to determine if notes should be initialized.
- * @returns A puzzle object that is readable by the Sudoku component.
- */
-export const returnGameOfDifficulty = (
-  difficulty: GameDifficulty | "dev",
-  initializeNotes: boolean,
-): SudokuObjectProps => {
-  const puzzles = returnPuzzleOfDifficulty(difficulty);
-  if (difficulty === "dev") {
-    difficulty = "novice";
-  }
-  return convertPuzzleToSudokuObject(puzzles, difficulty, initializeNotes);
-};
