+++ sudokuru/app/Data/drills/obvious_single_drills.ts
@@ -1,0 +1,3 @@
+export const OBVIOUS_SINGLE_DRILLS: string[] = [
+  "001000006902030040800402000000067900600100008050304000069000000080700020000003000",
+];
